+++ CHANGELOG.md
@@ -18,6 +18,7 @@
  - Update documentation for PositionComponent
  - Adding Component#onLoad
  - Moving size to Game instead of BaseGame
+ - Fix bug with ConcurrentModificationError on add in onMount
 
 ## 1.0.0-rc2
  - Improve IsometricTileMap and Spritesheet classes